Refactor out utils
MetNet is giving NaN when training, so working on figuring out why
