+++ satflow/data/datasets.py
@@ -1,6 +1,5 @@
-import datetime
 import glob
-from typing import Any, Dict, Iterator, List, Optional, Sequence, Type, Union, Tuple
+from typing import Iterator, List, Type, Union, Tuple
 
 import albumentations as A
 import numpy as np
@@ -9,11 +8,18 @@ import webdataset as wds
 from torch.utils.data.dataset import T_co
 import logging
 import pickle
-import io
 import random
 import os
 from satflow.data.utils.normalization import metnet_normalization, standard_normalization
-
+from satflow.data.utils.utils import (
+    create_time_layer,
+    load_np,
+    binarize_mask,
+    create_pixel_coord_layers,
+    check_channels,
+    crop_center,
+    load_config,
+)
 
 logger = logging.getLogger("satflow.dataset")
 logger.setLevel(logging.WARN)
@@ -33,29 +39,6 @@ def get_dataset(name: str) -> Type[thd.IterableDataset]:
     global REGISTERED_DATASET_CLASSES
     assert name in REGISTERED_DATASET_CLASSES, f"available class: {REGISTERED_DATASET_CLASSES}"
     return REGISTERED_DATASET_CLASSES[name]
-
-
-def create_time_layer(dt: datetime.datetime, shape):
-    """Create 3 layer for current time of observation"""
-    month = dt.month / 12
-    day = dt.day / 31
-    hour = dt.hour / 24
-    # minute = dt.minute / 60
-    return np.stack([np.full(shape, month), np.full(shape, day), np.full(shape, hour)], axis=-1)
-
-
-def load_np(data):
-    import numpy.lib.format
-
-    stream = io.BytesIO(data)
-    return numpy.lib.format.read_array(stream)
-
-
-def binarize_mask(mask):
-    """Binarize mask, taking max value as the data, and setting everything else to 0"""
-    tmp_mask = np.zeros_like(mask)
-    tmp_mask[np.isclose(np.round(mask), 2)] = 1
-    return tmp_mask
 
 
 # Taken from training set
@@ -91,75 +74,6 @@ MSG_STD = {
 
 TOPO_MEAN = 224.3065682349895
 TOPO_STD = 441.7514422990341
-
-
-def create_pixel_coord_layers(x_dim: int, y_dim: int, with_r: bool = False) -> np.ndarray:
-    """
-    Creates Coord layer for CoordConv model
-
-    :param x_dim: size of x dimension for output
-    :param y_dim: size of y dimension for output
-    :param with_r: Whether to include polar coordinates from center
-    :return: (2, x_dim, y_dim) or (3, x_dim, y_dim) array of the pixel coordinates
-    """
-    xx_ones = np.ones([1, x_dim], dtype=np.int32)
-    xx_ones = np.expand_dims(xx_ones, -1)
-
-    xx_range = np.expand_dims(np.arange(x_dim), 0)
-    xx_range = np.expand_dims(xx_range, 1)
-
-    xx_channel = np.matmul(xx_ones, xx_range)
-    xx_channel = np.expand_dims(xx_channel, -1)
-
-    yy_ones = np.ones([1, y_dim], dtype=np.int32)
-    yy_ones = np.expand_dims(yy_ones, 1)
-
-    yy_range = np.expand_dims(np.arange(y_dim), 0)
-    yy_range = np.expand_dims(yy_range, -1)
-
-    yy_channel = np.matmul(yy_range, yy_ones)
-    yy_channel = np.expand_dims(yy_channel, -1)
-
-    xx_channel = xx_channel.astype("float32") / (x_dim - 1)
-    yy_channel = yy_channel.astype("float32") / (y_dim - 1)
-
-    xx_channel = xx_channel * 2 - 1
-    yy_channel = yy_channel * 2 - 1
-    ret = np.stack([xx_channel, yy_channel], axis=0)
-
-    if with_r:
-        rr = np.sqrt(np.square(xx_channel - 0.5) + np.square(yy_channel - 0.5))
-        ret = np.concatenate([ret, np.expand_dims(rr, axis=0)], axis=0)
-    ret = np.moveaxis(ret, [1], [0])
-    return ret
-
-
-def check_channels(config: dict) -> int:
-    """
-    Checks the number of channels needed per timestep, to use for preallocating the numpy array
-    Is not the same as the one for training, as that includes the number of channels after the array is partly
-    flattened
-    Args:
-        config:
-
-    Returns:
-
-    """
-    channels = len(config.get("bands", []))
-    channels = channels + 1 if config.get("use_mask", False) else channels
-    channels = (
-        channels + 3
-        if config.get("use_time", False) and not config.get("time_aux", False)
-        else channels
-    )
-    # if config.get("time_as_channels", False):
-    # Calc number of channels + inital ones
-    #    channels = channels * (config["num_timesteps"] + 1)
-    channels = channels + 1 if config.get("use_topo", False) else channels
-    channels = channels + 3 if config.get("use_latlon", False) else channels
-    channels = channels + 2 if config.get("add_pixel_coords", False) else channels
-    channels = channels + 1 if config.get("add_polar_coords", False) else channels
-    return channels
 
 
 @register_dataset
@@ -732,14 +646,6 @@ class OpticalFlowDataset(SatFlowDataset):
                         yield prev_mask, mask, target_mask, image, prev_image
 
 
-def crop_center(img: np.ndarray, cropx: int, cropy: int) -> np.ndarray:
-    """Crops center of image through timestack, fails if all the images are concatenated as channels"""
-    t, c, y, x = img.shape
-    startx = x // 2 - (cropx // 2)
-    starty = y // 2 - (cropy // 2)
-    return img[:, :, starty : starty + cropy, startx : startx + cropx]
-
-
 class FileDataset(thd.Dataset):
     def __getitem__(self, index) -> T_co:
         try:
@@ -772,11 +678,6 @@ if __name__ == "__main__":
     # NIR1.6, VIS0.8 and VIS0.6 RGB for near normal view
     import numpy as np
     import webdataset as wds
-    import yaml
-
-    def load_config(config_file):
-        with open(config_file, "r") as cfg:
-            return yaml.load(cfg, Loader=yaml.FullLoader)["config"]
 
     dataset = wds.WebDataset("../../datasets/satflow-test.tar")
     config = load_config("../tests/configs/satflow.yaml")
